add api module to separate fetch code
